Show warning in case the bookmarks file doesn't exist.

+++ main.ps1
@@ -1,9 +1,10 @@
 
 # Modify this - WHERE TO SAVE
 $targetSaveLocation = 'C:\temp\save'
+
+
 $pathChromeUserData = "$env:LOCALAPPDATA\Google\Chrome\User Data"
-$profiles = @{}
-                          
+$profiles = @{}                          
 
 $json = Get-Content -Raw "$pathChromeUserData\Local State" | ConvertFrom-Json
 $jsonFragment = $json.profile
@@ -17,11 +18,13 @@ foreach ($obj in $jsonFragment) {
   }
 }
 
-foreach ($p in $profiles.keys) {    
-  $newDirName = "${p} ($($profiles.$p))"  
-  mkdir "$targetSaveLocation\$newDirName"
-  Copy-Item "$pathChromeUserData\${p}\Bookmarks" "$targetSaveLocation\$newDirName"  
+foreach ($p in $profiles.keys) {   
+  if (Test-Path "$pathChromeUserData\${p}\Bookmarks") {
+    $newDirName = "${p} - $($profiles.$p)"
+    mkdir "$targetSaveLocation\$newDirName"
+    Copy-Item "$pathChromeUserData\${p}\Bookmarks" "$targetSaveLocation\$newDirName"
+  }  
+  else {
+    Write-Warning "No Bookmarks file for $($profiles.$p) Chrome profile."
+  }
 }
-
-
-
